Changed cases where timing was being passed as a float to as an int. Makes no sense to have a partial-millisecond delta time accuracy. Also moved the updating for the Character's sprite and interpolation out of Draw and into HandleUpdate, which, along with some rounding, seems to fix the top-down jittering.

+++ netgore/trunk/DemoGame.MapEditor/MapEditorCharacter.cs
@@ -47,7 +47,7 @@ namespace DemoGame.MapEditor
         /// Perform pre-collision velocity and position updating.
         /// </summary>
         /// <param name="deltaTime">The amount of that that has elapsed time since last update.</param>
-        public override void UpdateVelocity(float deltaTime)
+        public override void UpdateVelocity(int deltaTime)
         {
             SetVelocityRaw(Vector2.Zero);
         }
