+++ netgore/trunk/NetGore/World/Entities/DynamicEntity.cs
@@ -342,7 +342,7 @@ namespace NetGore
         /// </summary>
         /// <param name="imap">The map the <see cref="Entity"/> is on.</param>
         /// <param name="deltaTime">The amount of time (in milliseconds) that has elapsed since the last update.</param>
-        protected override void HandleUpdate(IMap imap, float deltaTime)
+        protected override void HandleUpdate(IMap imap, int deltaTime)
         {
             base.HandleUpdate(imap, deltaTime);
